Save point. Creating filter catalog complete.

+++ online_market_data/src/main/java/com/em/online_market_data/products/sources/mappers/MapProductDTOinDBO.kt
@@ -29,7 +29,7 @@ fun mapProductDTOInDBO(productDTO: ProductDTO, productImages:  Map<String, Produ
 
 private fun mapProductPriceDTOInDBO(productPriceDTO: PriceDTO): PriceDBO {
     return PriceDBO(
-        price = productPriceDTO.price,
+        price = productPriceDTO.price.toInt(),
         discount = productPriceDTO.discount,
         unit = productPriceDTO.unit,
         priceWithDiscount = productPriceDTO.priceWithDiscount
